+++ examples/aws_provider/test/fixtures/us_east_1/main.tf
@@ -1,7 +1,3 @@
-terraform {
-  required_version = "~> 0.10.2"
-}
-
 variable "public_key_material" {
   description = "The public key material to use for SSH authentication with the instances"
   type        = "string"
@@ -23,6 +19,11 @@ output "reachable_other_host_public_ip" {
 output "security_group" {
   description = "The name of the security group"
   value       = "${module.example.security_group}"
+}
+
+output "terraform_state" {
+  description = "The path to the backend state file"
+  value       = "${module.example.terraform_state}"
 }
 
 output "test_target_public_dns" {