+++ spec/lib/kitchen/provisioner/terraform_spec.rb
@@ -14,41 +14,99 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+require "dry/monads"
+require "kitchen"
+require "kitchen/driver/terraform"
 require "kitchen/provisioner/terraform"
-require "support/kitchen/driver/terraform_context"
-require "support/kitchen/instance_context"
 require "support/kitchen/terraform/configurable_examples"
 
 ::RSpec
   .describe ::Kitchen::Provisioner::Terraform do
-    include_context "Kitchen::Instance initialized"
-
     let :described_instance do
-      provisioner
+      described_class.new({})
     end
 
     it_behaves_like "Kitchen::Terraform::Configurable"
 
     describe "#call" do
+      let :driver do
+        ::Kitchen::Driver::Terraform.new
+      end
+
+      let :kitchen_instance do
+        ::Kitchen::Instance
+          .new(
+            driver: driver,
+            logger: ::Kitchen::Logger.new,
+            platform: ::Kitchen::Platform.new(name: "test-platform"),
+            provisioner: described_instance,
+            state_file:
+              ::Kitchen::StateFile
+                .new(
+                  "/kitchen/root",
+                  "test-suite-test-platform"
+                ),
+            suite: ::Kitchen::Suite.new(name: "test-suite"),
+            transport: ::Kitchen::Transport::Base.new,
+            verifier: ::Kitchen::Verifier::Base.new
+          )
+      end
+
+      let :kitchen_state do
+        {}
+      end
+
+      before do
+        described_instance.finalize_config! kitchen_instance
+      end
+
       subject do
         lambda do
-          described_instance.call instance_double ::Object
+          described_instance.call kitchen_state
         end
       end
 
-      context "when the driver create action is a failure" do
-        include_context "Kitchen::Driver::Terraform#create failure"
+      describe "error handling" do
+        context "when the driver create action is a failure" do
+          before do
+            allow(driver).to receive(:apply).and_return ::Dry::Monads.Left "mocked Driver#create failure"
+          end
 
-        it "raises an action failed error" do
-          is_expected.to raise_error ::Kitchen::ActionFailed
+          it do
+            is_expected
+              .to(
+                raise_error(
+                  ::Kitchen::ActionFailed,
+                  "mocked Driver#create failure"
+                )
+              )
+          end
+        end
+
+        context "when the driver create action is a success" do
+          before do
+            allow(driver).to receive(:apply).and_return ::Dry::Monads.Right "mocked Driver#create output"
+          end
+
+          it do
+            is_expected.to_not raise_error
+          end
         end
       end
 
-      context "when the driver create action is a success" do
-        include_context "Kitchen::Driver::Terraform#create success"
+      describe "Test Kitchen state manipulation" do
+        before do
+          allow(driver).to receive(:apply).and_return ::Dry::Monads.Right "mocked Driver#create output"
 
-        it "does not raise an error" do
-          is_expected.to_not raise_error
+          described_instance.call kitchen_state
+        end
+
+        subject do
+          kitchen_state
+        end
+
+        it do
+          is_expected.to include kitchen_terraform_output: "mocked Driver#create output"
         end
       end
     end