+++ spec/lib/kitchen/verifier/terraform/configure_inspec_runner_attributes_spec.rb
@@ -14,84 +14,72 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-require "kitchen/terraform/client"
 require "kitchen/verifier/terraform/configure_inspec_runner_attributes"
 require "support/dry/monads/either_matchers"
-require "support/kitchen/driver/terraform_context"
 
-::RSpec.describe ::Kitchen::Verifier::Terraform::ConfigureInspecRunnerAttributes do
-  describe ".call" do
-    include_context "Kitchen::Driver::Terraform finalized instance"
-
-    let :group do
-      {}
-    end
-
-    subject do
-      described_class.call driver: driver,
-                           group: group,
-                           terraform_state: "terraform_state"
-    end
-
-    context "when the Terraform output command results in failure" do
-      include_context "Kitchen::Driver::Terraform#output failure"
-
-      it do
-        is_expected
-          .to result_in_failure.with_the_value /configuring Inspec::Runner attributes failed.*terraform output/m
-      end
-    end
-
-    context "when the value of the Terraform output command result is unexpected" do
-      include_context(
-        "Kitchen::Driver::Terraform#output success",
-        output: ::JSON.generate("name" => {"unexpected" => "value"})
-      )
-
-      it do
-        is_expected.to result_in_failure.with_the_value /configuring Inspec::Runner attributes failed.*\"value\"/m
-      end
-    end
-
-    context "when the group attribute output names do not match the value of the Terraform output command result" do
-      include_context(
-        "Kitchen::Driver::Terraform#output success",
-        output: ::JSON.generate("output_name" => {"value" => "output_name value"})
-      )
-
+::RSpec
+  .describe ::Kitchen::Verifier::Terraform::ConfigureInspecRunnerAttributes do
+    describe ".call" do
       let :group do
-        {
-          attributes: {
-            attribute_name: "not_output_name"
-          }
-        }
+        {}
       end
 
-      it do
-        is_expected
-          .to result_in_failure.with_the_value /configuring Inspec::Runner attributes failed.*\"not_output_name\"/m
-      end
-    end
-
-    context "when the group attribute output names match the value of the Terraform output command result" do
-      include_context(
-        "Kitchen::Driver::Terraform#output success",
-        output: ::JSON.generate("output_name" => {"value" => "output_name value"})
-      )
-
-      let :group do
-        {
-          attributes: {
-            attribute_name: "output_name"
-          }
-        }
+      subject do
+        described_class
+          .call(
+            group: group,
+            output: output
+          )
       end
 
-      it do
-        is_expected.to result_in_success.with_the_value "attribute_name" => "output_name value",
-                                                        "output_name" => "output_name value",
-                                                        "terraform_state" => "terraform_state"
+      context "when the value of the Terraform output command result is unexpected" do
+        let :output do
+          {"name" => {"unexpected" => "value"}}
+        end
+
+        it do
+          is_expected
+            .to result_in_failure.with_the_value /Configuring InSpec runner attributes resulted in failure: .*\"value\"/
+        end
+      end
+
+      context "when the group attribute output names do not match the value of the Terraform output command result" do
+        let :group do
+          {attributes: {attribute_name: "not_output_name"}}
+        end
+
+        let :output do
+          {"output_name" => {"value" => "output_name value"}}
+        end
+
+        it do
+          is_expected
+            .to(
+              result_in_failure
+                .with_the_value(/Configuring InSpec runner attributes resulted in failure: .*\"not_output_name\"/)
+            )
+        end
+      end
+
+      context "when the group attribute output names match the value of the Terraform output command result" do
+        let :group do
+          {attributes: {attribute_name: "output_name"}}
+        end
+
+        let :output do
+          {"output_name" => {"value" => "output_name value"}}
+        end
+
+        it do
+          is_expected
+            .to(
+              result_in_success
+                .with_the_value(
+                  "attribute_name" => "output_name value",
+                  "output_name" => "output_name value"
+                )
+            )
+        end
       end
     end
   end
-end